// web/src/lib/api/types/sanity.types.ts
export type Track = {
  _id: string;
  _type: "track";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  alias?: {
    _ref: string;
    _type: "reference";
    _weak?: boolean;
    [internalGroqTypeReferenceTo]?: "alias";
  };
  title?: string;
  description?: Array<{
    children?: Array<{
      marks?: Array<string>;
      text?: string;
      _type: "span";
      _key: string;
    }>;
    style?: "normal" | "h1" | "h2" | "h3" | "h4" | "h5" | "h6" | "blockquote";
    listItem?: "bullet" | "number";
    markDefs?: Array<{
      href?: string;
      _type: "link";
      _key: string;
    }>;
    level?: number;
    _type: "block";
    _key: string;
  } | {
    _key: string;
  } & Instrument>;
  coverart?: {
    asset?: {
      _ref: string;
      _type: "reference";
      _weak?: boolean;
      [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
    };
    media?: unknown;
    hotspot?: SanityImageHotspot;
    crop?: SanityImageCrop;
    _type: "image";
  };
  duration?: string;
  year?: string;
  genre?: Array<{
    _ref: string;
    _type: "reference";
    _weak?: boolean;
    _key: string;
    [internalGroqTypeReferenceTo]?: "genre";
  }>;
  type?: "Original" | "Remix";
  slug?: Slug;
  playable?: {
    provider?: "Spotify" | "SoundCloud";
    uri?: string;
  };
};

export type Slug = {
  _type: "slug";
  current?: string;
  source?: string;
};

export type SanityImageCrop = {
  _type: "sanity.imageCrop";
  top?: number;
  bottom?: number;
  left?: number;
  right?: number;
};

export type SanityImageHotspot = {
  _type: "sanity.imageHotspot";
  x?: number;
  y?: number;
  height?: number;
  width?: number;
};

export type Platform = {
  _id: string;
  _type: "platform";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  name?: string;
  url?: string;
};

export type Instrument = {
  _type: "instrument";
  manufacturer?: string;
  model?: string;
  type?: "Synthesizer" | "Drum machine" | "Sampler" | "Effect" | "Other";
};
// ...
export type Alias = {
  _id: string;
  _type: "alias";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  name?: string;
  slug?: Slug;
  genre?: string;
  url?: Array<{
    _key: string;
  } & Platform>;
};
// ...
export declare const internalGroqTypeReferenceTo: unique symbol;